finalizado a verificacao de texturas e cores do mapa, com norminette
